pyproject.toml
[project]
name = "uniquedeep"
version = "0.1.0"
description = "使用 LangChain 构建的具有 Skills 特性的 Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "rich>=14.2.0",
    "prompt-toolkit>=3.0.0",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    # LangChain 1.0 Skills Agent 依赖
    "langchain>=1.0.0",
    "langchain-anthropic>=1.0.0",
    "langchain-deepseek>=1.0.0",
    "langchain-google-genai>=1.0.0",
    "langchain-openai>=1.0.0",
    "langgraph>=1.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.scripts]
uniquedeep = "uniquedeep.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/uniquedeep"]

[tool.uv.sources]
uniquedeep-agent = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
